Refactored error messages based on review suggestions
